Move commonly used attributes functions to utils
